Add bd sqlite

+++ app/shemas.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel, Field, field_validator
@@ -5,12 +6,12 @@ from pydantic import BaseModel, Field, field_validator
 
 class OperationsRequest(BaseModel):
     wallet_name: str = Field(..., max_length=100)
-    amount: float
+    amount: Decimal
     description: Optional[str] = Field(None, max_length=255)
 
     # Валидация суммы - она должна быть положительной
     @field_validator('amount')
-    def amount_must_be_positive(cls, value: float) -> float:
+    def amount_must_be_positive(cls, value: Decimal) -> Decimal:
         if value <= 0:
             raise ValueError('Amount must be greater than zero')
         return value
@@ -25,7 +26,7 @@ class OperationsRequest(BaseModel):
 
 class CreateWalletRequest(BaseModel):
     name: str = Field(..., max_length=100)
-    initial_balance: float = 0
+    initial_balance: Decimal = 0
 
     @field_validator('name')
     def name_not_empty(cls, value: str) -> str:
@@ -35,7 +36,7 @@ class CreateWalletRequest(BaseModel):
         return value
 
     @field_validator('initial_balance')
-    def balance_not_negative(cls, value: float) -> float:
+    def balance_not_negative(cls, value: Decimal) -> Decimal:
         if value < 0:
             raise ValueError('Initial balance cannot be negative')
         return value
